check the configuration for references to remote files where only local files can be referenced
